% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lags.R
\name{calc.lags}
\alias{calc.lags}
\title{Calculate (multiple) Lags of a time series}
\usage{
calc.lags(ts, lags)
}
\arguments{
\item{ts}{A vector of a time series.}

\item{lags}{A vector of lags (positive or negative).}
}
\value{
A \code{data.frame} containing the ts at different lags of the same length as the original ts. NAs are used to fill in the missing values resulting from the shift.
}
\description{
Function to calculate lagged versions of time-series.
}
