% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mean.over.Month.R
\name{calc.mean.over.Month}
\alias{calc.mean.over.Month}
\title{Seasonal mean for spatio-temporal field}
\usage{
calc.mean.over.Month(raster, time, month, shiftYear = TRUE, verbose = T)
}
\arguments{
\item{raster}{A raster of class \code{RasterBrick} ('raster'-pkg) or \code{SpatRaster} ('terra'-pkg) with each layer corresponding to a time step.}

\item{time}{A vector of dates corresponding to each layer in the raster \code{brick}.}

\item{month}{A vector of numbers between 1 and 12 denoting the month over which to calculate the average (e.g. 3,4,5). Only subsequent month are allowed (e.g \code{c(12,1,2)} works; \code{c(1,3,5)} does not!).}

\item{shiftYear}{A logical value. How to handle averages over subsequent month but different years (e.g. Dec.-Jan.-Feb.). If \code{TRUE} the time returned corresponds to the older year (e.g. 1989), otherwise to the younger one (1990).}

\item{verbose}{A logical value. Print output to console?}
}
\value{
A raster of class \code{RasterBrick} ('raster'-pkg) or \code{SpatRaster} ('terra'-pkg) containing the avg. spatial fields per specified monthly period per year.
}
\description{
Function to calculate the mean over subsequent month in a raster brick.
}
\examples{
## For an artificial toy dataset
library(raster)
library(terra)
r = raster::raster(extent(c(-5,5,40,60)),
           res = c(0.1,0.1))
date = seq(as.Date("1990-01-01"),
           as.Date("2000-12-31"),by = "month")
br = raster::brick(lapply(seq_along(date), function(i) setValues(r,runif(ncell(r)))))
calc.mean.over.Month(raster = br, time = date,month = c(12,1,2),shiftYear = TRUE)
calc.mean.over.Month(raster = br, time = date,month = c(12,1,2),shiftYear = FALSE)
calc.mean.over.Month(raster = terra::rast(br), time = date,month = c(12,1,2),shiftYear = FALSE)
calc.mean.over.Month(raster = terra::rast(br), time = date,month = 1:12,shiftYear = FALSE)
## For SST of the North Sea
data(sst.ahoi)
SST.winter = calc.mean.over.Month(raster = sst.ahoi,
                                  time = as.Date(sub("X","",names(sst.ahoi)),format = "\%Y.\%m.\%d"),
                                  month = c(12,1,2),shiftYear = TRUE)
plot(SST.winter,1:6)

}
