% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF.Monte.Carlo.R
\name{EOF.Monte.Carlo}
\alias{EOF.Monte.Carlo}
\title{Finds Nr. of PCs to retain by means of Monte-Carlo analysis}
\usage{
EOF.Monte.Carlo(
  raster,
  center = TRUE,
  scale = FALSE,
  size = 100,
  CI.size = 1000,
  plot = TRUE
)
}
\arguments{
\item{raster}{A raster brick from which to calculate the spatial-PCA.}

\item{center}{If the mean should be removed prior to analysis.}

\item{scale}{If each grid point should be scaled by its variance.}

\item{size}{Nr. of permutations to generate for the Null-model.}

\item{CI.size}{Nr. of bootstrap samples for calculating confidence-intervals.}

\item{plot}{A logical value, indicating if the eigenvalues of the Null-model, PCA-analysis and CIs are plotted.}
}
\value{
A \code{list} with the following elements:
\itemize{
\item eigen.data - Eigenvalues of the original dataset
\item eigen.data.bootstrap - Eigenvalues of the bootstrapped dataset
\item eigen0 - Eigenvalues of the Null-model
\item Nr.PCs - the number of PCs to retain based on comparing the eigenvalues of the Null-model with the bootstrapped eigenvalues.
}
}
\description{
A form of Horns parallel analysis to compare the eigenvalues of a PCA with those of a null-model (randomly permuted colums).
}
\details{
...
}
\references{
...
}
