% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldAnomaly.to.original.R
\name{fieldAnomaly.to.original}
\alias{fieldAnomaly.to.original}
\title{Add back field anomaly to a spatio-temporal field}
\usage{
fieldAnomaly.to.original(rst1, time, level = "month")
}
\arguments{
\item{rst1}{A raster either a \code{RasterBrick} from the raster package or \code{SpatRaster} from terra with each layer corresponding to a time step.}

\item{time}{A vector of the class "Date", "POSIXct", or "POSIXlt" corresponding to each layer in the raster.}

\item{level}{A character string. The temporal resolution over which to calculate the anomalies. Either \code{julian} / \code{day} for daily or \code{month} for monthly anomalies.
Will consider "Day of the year" from as.POSIXlt(x)$yday for \code{julian} and  unique format(x,"\%m\%d") strings for \code{day}.}
}
\value{
A raster either of form \code{RasterBrick} or \code{SpatRaster} containing the original field.
}
\description{
The counterpart of the \code{fieldAnomaly.raster(...)} function to add back the anomalies and get back the original field.
}
\details{
The field anomalies are calculated by subtracting the daily/monthly means from each spatial location of a field.
Therefore in order to get back to the original field the mean on each location is added back to the anomalies.
}
\examples{
library(raster)
data("sst.ahoi")

# calculate anomalies
time.ahoi = as.Date(sub("X","",names(sst.ahoi)),format = "\%Y.\%m.\%d")
sst.anom = fieldAnomaly.raster(sst.ahoi,
                              time = time.ahoi,
                              level = "month")
# back calculate the original field
original.field = fieldAnomaly.to.original(sst.anom,time = time.ahoi,level = "month")

# check if they are the same
all(cellStats(original.field,mean) -cellStats(sst.ahoi,mean)== 0)
# correct
}
\references{
Taylor, M. (2017). sinkr: Collection of functions with emphasis in multivariate data analysis.
R package version 0.6. \href{https://github.com/marchtaylor/sinkr}{https://github.com/marchtaylor/sinkr}
}
