% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessFuncs.R
\name{fitnessfuns}
\alias{fitnessfuns}
\alias{MO.fitness.func.RF.ranger}
\alias{MO.fitness.func.RF.regression}
\alias{MO.fitness.func.extraTrees.ranger}
\title{Random Forest & extreme randomized Trees multi-obj. fitness functions for use with NSGA-II feature selection}
\usage{
MO.fitness.func.RF.ranger(
  feat,
  x,
  y,
  seed,
  metric.func,
  folds = 3,
  reps = 20,
  only.obs.vs.pred = FALSE,
  return.scaled.metric = TRUE
)

MO.fitness.func.extraTrees.ranger(
  feat,
  x,
  y,
  seed,
  metric.func,
  folds = 3,
  reps = 20,
  only.obs.vs.pred = FALSE,
  return.scaled.metric = TRUE
)
}
\arguments{
\item{feat}{The binary coded feature vector that denotes which gene (feature) is on (1) or off (0).}

\item{x}{The dataframe containing all the features to run feature selection on.}

\item{y}{Numeric vector of the response variable.}

\item{seed}{The setting for the seed.}

\item{metric.func}{Slot for the metric function to be included. The format is metric.func(data) see \link[marmalaid]{metricfuns} for details.}

\item{folds}{Number of folds (k) for n-repeated k-fold crossvalidation.}

\item{reps}{Number of repetitions (n) for n-repeated k-fold crossvalidation.}

\item{only.obs.vs.pred}{A logical value. Setting to return only observations vs. predictions without any metric-function being considered.}

\item{return.scaled.metric}{A logical value. Setting if the metric should be scaled by a naive forecast (mean of the training data).}
}
\description{
Various multi-obj. fitness functions for use with NSGA-II feature selection
}
\details{
\code{MO.fitness.func.RF.ranger}  - Fitness function for the "randomForest" - algorithm from the \link[ranger]{ranger} package.
\cr \code{MO.fitness.func.extraTrees.ranger}  - Fitness function for the "extraTrees" - algorithm from the \link[ranger]{ranger} package.
}
\note{
\code{fitnessfuns} is a generic name for the functions documented.
}
