% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.list.R
\name{flatten.list}
\alias{flatten.list}
\title{Function to flatten a deeply nested list to a shallow one}
\usage{
flatten.list(x, verbose = TRUE)
}
\arguments{
\item{x}{A deeply nested \code{list}.}

\item{verbose}{Toggle printing of different layer nestings to console on or off if function steps through them.}
}
\value{
A shallow \code{list}
}
\description{
Function to flatten a deeply nested list to a shallow one.
}
\details{
A function that flattens a deeply nested list and returns a shallow flattened one with only one level of nesting.
Be aware that the single arguments are returned somewhat out of order, since the degree of nesting can vary between elements.
}
\examples{
# For an artificial toy dataset
lst = list(a = list(a = 1:10,b = list(1,2,3),c = list(a = list(1:100,NA,list(rep(NA,100))))),
           b = list(data.frame(id = 1:4,val = c(2,NA,3,5)),
                    list(letters,data.frame(m = month.abb,no = 1:length(month.abb)))))
lst_unnested = flatten.list(lst)
print(lst_unnested)

}
