% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.metrics.regression_NSGAII.R
\name{Model.metrics.regression_NSGAII}
\alias{Model.metrics.regression_NSGAII}
\title{Calculates Model metrics (RMSE,MAE,Pseudo-R2) for output of feature selection NSGA-II}
\usage{
Model.metrics.regression_NSGAII(NSGA.II.obj, fitness.func)
}
\arguments{
\item{NSGA.II.obj}{The output object from the NSGA-II feature selection.}

\item{fitness.func}{A function used as fitness function in the feature selection process.}
}
\value{
A \code{data.frame} specifying the solutions within the pareto front, with the following columns:
\itemize{
\item ID - the index in 'NSGA.II.obj$pareto.varnames' to which each solution corresponds to
\item metric - the metric used to evaluate the fitness of the solution.
\item params - a number between 0 and 1 representing the scaled number of parameters used for feature selection
\item nr.of.params - An integer. Specifies the number of covariates used in the model.
\item dist.to.CoordOrigin - distance of the solution within the pareto front to the coordinate origin (0,0)
\item RMSE - the RMSE (with the same level of Cross validation, which was used during the feature selection)
\item PseudoR2 - the pseudo R2 (with the same level of Cross validation, which was used during the feature selection)
\item MAE - the MAE (with the same level of Cross validation, which was used during the feature selection)
}
}
\description{
Function to calculate Model metrics (RMSE,MAE,Pseudo-R2) for output of feature selection NSGA-II
}
