% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.merge.R
\name{mult.merge}
\alias{mult.merge}
\title{Merge multiple tables}
\usage{
mult.merge(fileA, listFileB, by, by.firstCol = FALSE)
}
\arguments{
\item{fileA}{A string denoting the name of a data.frame}

\item{listFileB}{A character vector denoting data.frames or lists containg data.frames, which are merged with fileA.}

\item{by}{Variable name by which to merge.}

\item{by.firstCol}{A logical value. If merging should be done by the first column in each data.frame regardless of the name.}
}
\value{
A merged \code{data.frame}.
}
\description{
Function to merge various data.frames/list of data.frames to one.
}
