% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eof.R
\name{plot_eof}
\alias{plot_eof}
\title{Function to plot the EOF pattern (spatial/temporal)}
\usage{
plot_eof(eof.out, n)
}
\arguments{
\item{eof.out}{An output object from the \code{spatial.PCA} function}

\item{n}{The number of corresponding EOF-pattern/PC-time series to show.}
}
\description{
This functions is a rather easy plotting function to see the spatial & temporal
representation of the first n calculated EOF modes.
Typically a maximum of 15 modes should not be exceeded.
}
