% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGA.II_wrapper.R
\name{plot_NSGAII}
\alias{plot_NSGAII}
\title{plot function for output of \code{NSGA.II_wrapper}}
\usage{
plot_NSGAII(output.NSGAII)
}
\arguments{
\item{output.NSGAII}{An obj. (list) containing results of the NSGA-II feature selection.}
}
\value{
A plot showing the evolution with the Hyper-volume Indicator as well as the Pareto-Front.
}
\description{
A summary plot for the NSGA-II feature selection.
}
