% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.DVIndex.R
\name{SOM.DVIndex}
\alias{SOM.DVIndex}
\title{Dynamic Validity Index (DVI) of Shen et al. (2005) for SOMs}
\usage{
SOM.DVIndex(list.SOM.obj, c.param = 1, control.for.empty.nodes = FALSE)
}
\arguments{
\item{list.SOM.obj}{A \code{list} of SOM objects of \code{class(kohonen)} with different map sizes to be evaluated with the DVI.}

\item{c.param}{Control parameter to weight the influence of within cluster density and in-between cluster seperability.}

\item{control.for.empty.nodes}{Logical value, denoting if empty clusters should be discarded in the calculation of the in-between distance (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item xdim - the y-dimension of the SOM mapping.
\item ydim - the x-dimension of the SOM mapping
\item size - the total number of clusters (product of xdim and ydim)
\item intra.ratio - A measure of the inner cluster density.
\item inter.ratio - A measure of the seperability between clusters.
\item DVI - The dynamic validity index, being a combination of intra - and inter ratio.
}
}
\description{
Function to calculate the DVI for SOMs with various grid-sizes.
}
\details{
This function implements the DVI of Shen et al. (2005) for SOMs, which was originally proposed for kmeans clustering. It identifies the optimal number of clusters to be a compromise between intra
compactness of clusters and inter-seperatedness. Clusters should have a small distance to members of their own cluster (intra-ratio) as well as having a larger distance to members of other clusters (inter-ratio).
The DVI is now the sum of both the normalised intra- and inter ratios (optional scaling with a correction parameter is possible) and has a minimum at the prefered number of clusters. Since larger map sizes of SOMs can lead to empty clusters
(nodes that have no datapoint attached to them), a correction with the option \code{control.for.empty.nodes = TRUE} is possible, that only considers clusters with data attached to them.
}
\examples{
library(raster)
data(sst.ahoi)
sst.anom = fieldAnomaly.raster(brick = sst.ahoi,
                            time = as.Date(sub("X","",names(sst.ahoi)),format = "\%Y.\%m.\%d"),
                             level = "month")
# calc. mean over summer season
sst.JJA = calc.mean.over.Month(sst.anom,
                             time = as.Date(sub("X","",names(sst.anom)),
                                            format = "\%Y.\%m.\%d"),
                             month = c(6,7,8),shiftYear = FALSE)

# grid sizes
grid.size = data.frame(expand.grid(1:5,1:5))
names(grid.size) = c("xdim","ydim")
SOMs.store = list()
for(i in 1:nrow(grid.size)){
 cat(i,"\n")
 # calculate spatial SOM
 SOMs.store[[i]] = spatial.SOM(x = sst.JJA,time = as.numeric(sub("X","",names(sst.JJA))),
                              plot = FALSE,seed = NULL,
                              parallel = c(parallel = FALSE,cores = NA),
                              reps = 1,return.all = FALSE,
                              grid = kohonen::somgrid(xdim = grid.size$xdim[i],
                                                      ydim = grid.size$ydim[i],
                                                      topo = "rectangular",
                                                      neighbourhood.fct = "bubble",
                                                      toroidal = FALSE),
                              mode = "online",rlen = 1000)
}
# calculate DVI
DVI.SST.JJA = SOM.DVIndex(list.SOM.obj = lapply(SOMs.store,function(x) x$SOM.out),
                         c.param = 1,
                         control.for.empty.nodes = TRUE)
DVI.SST.JJA = DVI.SST.JJA[order(DVI.SST.JJA$size),]
# plot DVI
par(mfrow = c(1,1))
plot(1:nrow(DVI.SST.JJA),DVI.SST.JJA$DVI,type = "l",xaxt = "n",
     xlab = " grid size",ylab  = "DVI",las = 1)
points(1:nrow(DVI.SST.JJA),DVI.SST.JJA$DVI)
abline(v = which(DVI.SST.JJA$DVI == min(DVI.SST.JJA$DVI,na.rm = TRUE)),
       col = "red",lty = 2)
axis(side = 1,at = 1:nrow(DVI.SST.JJA),labels = DVI.SST.JJA$size)
title("DVI",adj = 0)
}
\references{
Shen J, Chang SI, Lee ES, Deng Y, Brown SJ (2005) "Determination of cluster number in clustering microarray data." Appl Math Comput 169:1172–1185
}
