% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.map.qual.measures.R
\name{SOM.qualityMetrics}
\alias{SOM.qualityMetrics}
\alias{SOM.quant.error}
\alias{SOM.topo.error}
\title{Quality metrics for the map representation of a SOM}
\usage{
SOM.quant.error(SOM.obj)

SOM.topo.error(SOM.obj)
}
\arguments{
\item{SOM.obj}{An object of \code{class(kohonen)} produced via \code{\link[kohonen:supersom]{som}}.}
}
\value{
\subsection{\code{SOM.quant.error}}{

A \code{numeric} value for the QE
}

\subsection{\code{SOM.topo.error}}{

A named \code{list} with following elements: \itemize{
\item TE - a numeric value for the Topographic error
\item C.index - a numeric value for the combined metric of QE & TE error from Kaski & Lagus 1996
\item Mapping.of.data - A \code{data.frame} showing the best matching unit (BMU), the second best matching unit (SNDMU), the C.index and an integer value denoting if BMU and SNDMU are neighbours (1) or not (0) for every data point.
}
}
}
\description{
Various quality metrics to assess the quality of the SOM-mapping returned by \code{\link[kohonen:supersom]{som}}.
}
\details{
\code{SOM.quant.error}  -  Calculates the Quantisation error (QE) for the SOM mapping
\cr \code{SOM.topo.error}  -  Calculates the Topographic error (TE) (Kiviluoto 1996) and a combined index of QE & TE based on the work of Kaski & Lagus 1996
}
\note{
\code{SOM.qualityMetrics} is a generic name for the functions documented.
}
\examples{
data(sst.ahoi)
# calculate field anomaly
sst.anom = fieldAnomaly.raster(brick = sst.ahoi,
                            time = as.Date(sub("X","",names(sst.ahoi)),format = "\%Y.\%m.\%d"),
                            level = "month")
# calc. mean over summer season
sst.JJA = calc.mean.over.Month(sst.anom,
                             time = as.Date(sub("X","",names(sst.anom)),
                                             format = "\%Y.\%m.\%d"),
                              month = c(6,7,8),shiftYear = FALSE)
# calculate spatial SOM
SST.SOM = spatial.SOM(x = sst.JJA,time = as.numeric(sub("X","",names(sst.JJA))),
                    plot = FALSE,seed = 1234,
                    parallel = c(parallel = TRUE,cores = 2),
                    reps = 50,return.all = FALSE,
                    grid = kohonen::somgrid(xdim = 2,ydim = 3,
                                            topo = "rectangular",
                                            neighbourhood.fct = "bubble",
                                            toroidal = FALSE),
                    mode = "online",rlen = 1000)
# calculate QE
SOM.quant.error(SST.SOM$SOM.out)
# calculate TE & C-index
SOM.topo.error(SST.SOM$SOM.out)
}
\references{
Kiviluoto K (1996) "Topology preservation in self-organizing maps." In: Proceedings of International Conference on Neural Networks (ICNN’96). IEEE, p 294–299
\cr \cr
Kaski S, Lagus K (1996) "Comparing self-organizing maps." In: von der Malsburg C , von Seelen W , Vorbrüggen JC, Sendhoff B (eds) Artificial Neural Networks — ICANN 96. ICANN 1996. Lecture Notes in Computer Science, vol 1112. Springer, Berlin, Heidelberg.p 809–814
\cr \cr
Pötzlbauer G (2004) "Survey and comparison of quality measures for self-organising maps." In: Proc. 5th Workshop on Data Analysis (WDA 2004).p 67–82
}
