% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.PCA.R
\name{spatial.PCA}
\alias{spatial.PCA}
\title{EOF analysis (S-mode PCA) for a spatio-temporal field}
\usage{
spatial.PCA(
  x,
  center = TRUE,
  scale. = FALSE,
  spatial.extent = "North Sea",
  plot = FALSE,
  var.threshold = 0.95
)
}
\arguments{
\item{x}{Either a raster \code{brick} for scalar valued spatio-temporal fields (e.g. SST) with each layer corresponding to a time step or a two-element \code{list} with each element being a \code{raster} for vector valued fields (e.g. u & v of Currents).
If layer names contain a reference to a date it will be used for writing it out.}

\item{center}{A logical value indicating whether the variables should be centered around the mean.}

\item{scale.}{A logical value indicating whether to scale each gridcell by its variance, if \code{FALSE} the PCA is calculated on the covariance matrix, otherwise (\code{TRUE}) on the correlation matrix.}

\item{spatial.extent}{Region to which the spatial field should be cropped, either \code{NULL} for no cropping, an object of class \code{extent} or the character string \code{"North Sea"} to cut it to the extent(c(xmin = -7.01,xmax = 12.9,ymin = 48.9,ymax = 61)) corresponding to the North Sea region.}

\item{plot}{A logical value if the spatial EOF fields should be returned in a plot after calculation.}

\item{var.threshold}{A value, indicating up to which percentage of explained variance the spatial fields should be returned (default = 0.95 (aka 95\% variance explained)).}
}
\value{
A \code{list} with the following elements:
\itemize{
\item raster - A \code{raster} for scalar valued input fields or a \code{list} of \code{rasters} for vector valued fields. Contains the spatial-EOF fields.
\item sp.PCA - A list returned by \link[stats]{prcomp} containing the output of the PCA with
\cr \code{x} - containing the PC timeseries and
\cr \code{rotation} - the eigenvectors (aka EOF-weights per grid-point).
\item time - the time associated to the PCs
}
}
\description{
Wrapper function for \code{prcomp(...)} on a spatio-temporal field organised as raster
}
\examples{
# load AHOI SST data
library(raster)
data(sst.ahoi)
# perform EOF-analysis
SST.EOF = spatial.PCA(sst.ahoi,center = TRUE,scale. = FALSE,spatial.extent = NULL,plot = FALSE)
# plot first EOF with PC timeseries
layout(matrix(rep(c(1,2),each = 2),nrow = 2,ncol = 2,byrow = TRUE),heights = c(0.65,0.35))
par(mar = c(4,4,3,3))
image(SST.EOF$raster[[1]],las = 1,xlab = "lon",ylab = "lat")
maps::map(add = TRUE,fill = TRUE,col = "gray80");box()
title("EOF1",adj = 0)
plot(SST.EOF$time,SST.EOF$sp.PCA$x[,1],type = "l",las = 1,xlab = "time",ylab = "")
title("PC1",adj = 0)
}
