% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.SOM.R
\name{spatial.SOM}
\alias{spatial.SOM}
\title{Self Organising Map (SOM) for a spatio-temporal field}
\usage{
spatial.SOM(
  x,
  time,
  plot = TRUE,
  seed = NULL,
  parallel = list(parallel = FALSE, cores = NULL),
  reps = 100,
  return.all = FALSE,
  ...
)
}
\arguments{
\item{x}{Either an object of class \code{RasterBrick} from the 'raster'-package or a \code{SpatRaster} from the 'terra'-package with each layer
corresponding to a time step specified in the argument \code{time}.}

\item{time}{the time step associated with each layer of the raster}

\item{plot}{A logical value if a simple plot with the convergence of the SOM algorithm and the BMU (best matching unit) time series should be plotted.}

\item{seed}{Seed settings for reproducibility, either \code{numeric} or \code{NULL}.}

\item{parallel}{List, containing two elements: parallel - logical, if analysis needs to be run in parallel mode, cores - number of CPU-cores to run analysis on}

\item{reps}{Number of repetitions to calculate for the SOM}

\item{return.all}{Logical. If \code{TRUE} all SOMs (specified in 'reps' besides the 'best' one) are returned.}

\item{...}{additional arguments specified in \code{\link[kohonen:supersom]{som}}.}
}
\value{
A named \code{list} with the following elements:
\itemize{
\item SOM.out - The 'best' SOM mapping found in regards to both QE and TE. Object of \code{class(kohonen)}. See \code{\link[kohonen:supersom]{som}} for details.
\item SOM.quality - The quality of the mapping assessed with both QE and TE.
\item SOM.raster - \code{raster} of the spatial-SOM fields.
\item Freq.pattern - the frequency of occurence of the spatial pattern within the BMU time series
\item BMU.ts - a \code{data.frame} showing the time and the associated BMU
\item return.all.SOMs - if \code{return.all.SOMs} is \code{TRUE}, a list with all stored SOM objects, otherwise \code{NULL}.
}
}
\description{
Wrapper function for \code{\link[kohonen:supersom]{som}} to perform a S-mode SOM on a spatio-temporal field organised as raster
}
\examples{
library(raster)
data(sst.ahoi)
# calculate field anomaly
sst.anom = fieldAnomaly.raster(rst1 = sst.ahoi,
                            time = as.Date(sub("X","",names(sst.ahoi)),format = "\%Y.\%m.\%d"),
                            level = "month")
# calc. mean over summer season
sst.JJA = calc.mean.over.Month(sst.anom,
                             time = as.Date(sub("X","",names(sst.anom)),
                                             format = "\%Y.\%m.\%d"),
                              month = c(6,7,8),shiftYear = FALSE)
# calculate spatial SOM
SST.SOM = spatial.SOM(x = sst.JJA,time = as.numeric(sub("X","",names(sst.JJA))),
                    plot = FALSE,seed = 1234,
                    parallel = c(parallel = TRUE,cores = 2),
                    reps = 50,return.all = FALSE,
                    grid = kohonen::somgrid(xdim = 3,
                                            ydim = 3,
                                            topo = "rectangular",
                                            neighbourhood.fct = "bubble",
                                            toroidal = FALSE),
                    mode = "online",rlen = 1000)
plot(SST.SOM$SOM.raster)

}
