% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldAnomaly.raster.R
\name{fieldAnomaly.raster}
\alias{fieldAnomaly.raster}
\title{Field anomaly for a spatio-temporal field}
\usage{
fieldAnomaly.raster(brick, time, level = "month")
}
\arguments{
\item{brick}{A raster \code{brick} with each layer corresponding to a time step.}

\item{time}{A vector of the class "Date", "POSIXct", or "POSIXlt" corresponding to each layer in the raster \code{brick}.}

\item{level}{A character string. The temporal resolution over which to calculate the anomalies. Either \code{julian} / \code{day} for daily or \code{month} for monthly anomalies.
Will consider "Day of the year" from as.POSIXlt(x)$yday for \code{julian} and  unique format(x,"\%m\%d") strings for \code{day}.}
}
\value{
A raster \code{brick} containing the field anomalies.
}
\description{
Wrapper function for \code{fieldAnomaly(...)} from the package sinkr to calculate the daily/monthly field anomaly for a raster brick.
}
\details{
The field anomalies are calculated by subtracting the daily/monthly means from each spatial location of a field.
}
\examples{
# load AHOI SST data
library(raster)
data(sst.ahoi)
# calculate anomalies
sst.anom = fieldAnomaly.raster(sst.ahoi,
                               time = as.Date(sub("X","",names(sst.ahoi)),format = "\%Y.\%m.\%d"),
                               level = "month")
# plot for comparison
par(mfrow = c(1,2))
plot(sst.anom[[1]],las = 1)
title("SST-anomaly",adj = 0)
plot(sst.ahoi[[1]],las = 1)
title("SST-raw",adj = 0)
}
\references{
Taylor, M. (2017). sinkr: Collection of functions with emphasis in multivariate data analysis.
R package version 0.6. \href{https://github.com/marchtaylor/sinkr}{https://github.com/marchtaylor/sinkr}
}
