% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFuncs.R
\name{metricfuns}
\alias{metricfuns}
\alias{RRSE}
\alias{RMSE}
\alias{RMSPE}
\alias{MAE}
\alias{PseudoR2}
\title{Various metric functions for performance evaluation}
\usage{
RRSE(data)

RMSPE(data)

RMSE(data)

MAE(data)

PseudoR2(data)
}
\arguments{
\item{data}{A \code{dataframe} containing a column with numeric observations (\code{obs}) and the prediction (\code{pred}).}
}
\description{
Various metric functions to assess the fit of a statistical model in a regression context.
}
\details{
{\code{RMSE} calculates the "Root mean squared error".
\code{MAE} calculates the "Mean absolute error".
\code{PseudoR2} calculates a "R2"-like metric as cor(obs,pred)^2.
\code{RMSPE} calculates the "Root mean squared percentage error".
\code{RRSE} calculates the "Root relative squared error", which is the RMSE relative to the RMSE of a naive prediction by using the deviations to the mean.
}
}
\note{
\code{metricfuns} is a generic name for the functions documented.
}
