% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSGA.II_wrapper.R
\name{NSGA.II_wrapper}
\alias{NSGA.II_wrapper}
\title{NSGA-II Wrapper function to perform multi-objective (2) feature selection}
\usage{
NSGA.II_wrapper(
  x,
  y,
  seed,
  ga.input,
  memoisation = F,
  mutation.rate = 0.1,
  crossover.rate = 0.8,
  remove.overlap = T,
  pop.size = 50,
  offspring.size = NULL,
  max.iter = 150,
  initialize.equal = T,
  init.limit = NULL,
  stop.criterion = NULL,
  n.cores = NULL,
  allowZeroGenes = F,
  ref.point = c(1, 1)
)
}
\arguments{
\item{x}{A \code{data.frame} containing all features that should be tried in the feature selection process.}

\item{y}{The response variable.}

\item{seed}{Defines the seed used for the internal loop to make sure to get the same fit with the same features used.}

\item{ga.input}{A named \code{list} list(fitness.func, metric.func) containing the fitness function & the metric used to evaluate the fit (e.g. RMSE).}

\item{memoisation}{A logical value. If memoisation should be used (\code{TRUE}) or not (\code{FALSE}). If memoisation is used, function calls with the same input parameters (aka same features) are stored in a Cache and returned if solution is visited again.}

\item{mutation.rate}{Probability of a mutation (bitflip) within the feature vector.}

\item{crossover.rate}{Fraction of crossover between parent solutions.}

\item{remove.overlap}{A logical value. If overlapping solutions should be removed (see Nojima et al. 2005)?}

\item{pop.size}{Parent population size for the genetic algorithm.}

\item{offspring.size}{Offspring population size for the genetic algorithm.}

\item{max.iter}{Maximum number of iterations the algorithm should run.}

\item{initialize.equal}{A logical value. If initial population should be drawn at random (\code{FALSE}) or if it should contain individuals drawn uniformly along the number of features dimension (\code{TRUE}).}

\item{init.limit}{A number. Maximum number of genes to activate in the initial population. This is rather useful if the model fitting is slow for models with lots of features (so you want to limit it for computational purposes).}

\item{stop.criterion}{Maximum number of iterations without improvement of the solution.}

\item{n.cores}{Number of CPU-cores to parallelize the external loop (parallelize number of individuals in the population). If \code{NULL} the function will try all possible number of cores and chooses the fastest (the evaluation will however take some time, based on the machine).}

\item{allowZeroGenes}{A logical value. If individuals containing only zeros (no features) should be included in the search (\code{TRUE}) or not (\code{FALSE}).}

\item{ref.point}{Vector of length 2, defining the Reference point for the Hypervolume Indicator.}
}
\value{
A \code{list} with the following elements:
\itemize{
\item generation.fitness - \code{data.frame} of tracked fitness per iteration
\item generation.populations - A list of all individuals within the population per iteration
\item generation.offspring - A list of all offspring individuals per iteration
\item pareto.individuals - individuals of the pareto-optimal solution
\item pareto.varnames - Names of features included in the pareto-optimal solution
\item pareto.solution - Summary table of the final pareto-solution
\item algorithm.params - Input parameters and other summarized in a list to make output more reproducible
\item fitted.dataset - \code{list} of \code{x} and \code{y} used for model fitting
}
}
\description{
Performs Bi-objective Feature selection via a specified fitness function
}
\details{
This function implements the non-dominated sorting genetic algorithm (NSGA-II) feature selection for the bi-objective case. It simultanously minimizes the Nr. of features and
a measure of performance (e.g. RMSE) for a machine learning model (e.g. Random Forest).
}
